elbow_rehab/service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3

pydantic==2.9.0

google-cloud-bigquery==3.34.0
google-cloud-storage==3.1.0

firebase-admin==6.6.0

sqlalchemy==2.0.37
psycopg2-binary==2.9.10

pandas==2.2.3
dacite==1.9.2
toolz==1.0.0

pytest==8.4.0
black
ruff

ahrs==0.3.1
scipy==1.11.2
numpy==1.26.0